- Renamed macroscope to microscope
