+++ glowtracker/TestFunctions/testObjectDetection.py
@@ -14,7 +14,7 @@ currentdir = os.path.dirname(os.path.abspath(inspect.getfile(inspect.currentfram
 parentdir = os.path.dirname(currentdir)
 sys.path.insert(0, parentdir) 
 
-from Macroscope_macros import extractWorms, extractWormsDiff, extractWormsCMS
+from Microscope_macros import extractWorms, extractWormsDiff, extractWormsCMS
 
 single = True
 mode = 'CMS'
